+++ src/main/java/net/bizlogin/oauth/kakao/service/KakaoServiceImpl.java
@@ -1,20 +1,24 @@
 package net.bizlogin.oauth.kakao.service;
 
+import javax.annotation.PostConstruct;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.github.scribejava.apis.KakaoApi;
+import com.github.scribejava.apis.NaverApi;
 import com.github.scribejava.core.builder.ServiceBuilder;
 import com.github.scribejava.core.model.OAuth2Authorization;
 import com.github.scribejava.core.oauth.OAuth20Service;
 
-import net.bizlogin.oauth.AbstractOAuthService;
-
+/**
+ * 프로모터 카카오 로그인
+ *
+ */
 @Service
-public class KakaoServiceImpl extends AbstractOAuthService {
+public class KakaoServiceImpl {
 
 	private static final Logger logger = LoggerFactory.getLogger(KakaoServiceImpl.class);
 
@@ -23,21 +27,34 @@ public class KakaoServiceImpl extends AbstractOAuthService {
 
 	@Value("${bizlogin.kakao.callbackUrl}")
 	private String callbackUrl;
-	
+
+	private OAuth20Service oauthService;
+
+	/**
+	 * 카카오 oauthService 생성
+	 */
+	@PostConstruct
+	public void createService() {
+		this.oauthService = new ServiceBuilder(appKey).callback(callbackUrl).build(NaverApi.instance());
+	}
+
+	/**
+	 * 카카오 로그인 인증 url 생성
+	 * @param state
+	 * @return url
+	 */
+	public String getAuthorizationUrl(String state) {
+		String authorizationUrl = this.oauthService.getAuthorizationUrl(state);
+		return authorizationUrl;
+	}
+
+	/**
+	 * 카카오 로그인
+	 * @param auth {code, state}
+	 */
 	@Transactional(rollbackFor = { Exception.class }, readOnly = false)
 	public void signin(OAuth2Authorization auth) {
 		logger.debug(oauthService.toString());
-	}
-	
-	@Override
-	public OAuth20Service createService() {
-		OAuth20Service oauthService = new ServiceBuilder(appKey).callback(callbackUrl).build(KakaoApi.instance());
-		return oauthService;
-	}
-
-	@Override
-	public boolean isValidToken(String token) {
-		return true;
 	}
 
 }
